Final tweaks for data model
Basically I need the activity summary json file associated with
each FIT file. The json file contains high level summary data like
name, activityId, calories, etc. I can use the activityId as a tag
which is applied to the activity details. This decreases overall
storage needs and makes the json summaries the source of 'truth'
instead of both details and summaries containing activityIds and
names which was polluting the database.